Improve children comparison by extracting all text
As mentioned before, when comparing children as string it is easy to
check if it as changed or not, when using objects it becomes hard since
they might have the exact render data, but their object id change, thus
triggering a re-render.

So by extracting all text in the children nodes, and compare the text
now and before, we have a better idea of what as changed.
This doesn't cover style changes, as in font size or weight changes,
those can affect the height enough to change the component from static
to collapsible. I don't consider them relevant (at least for this stage).

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>