Setup code analysis tool
Currently, there are 3 tools checking the code:
- Codacy
- CodeFactor
- Coveralls

Codacy does the same of both CodeFactor and Coveralls.
In comparison to CodeFactor, it seems to have more flexibility and code
patterns and other features like code coverage or code duplication.
Regards Coveralls, this one only takes care of the code coverage and
because Codacy also does that I'm removing this one.

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>